refactor(docs-retrieval): extract bundled from content-fetcher
Move the `bundled:` content loaders out of content-fetcher.ts into a focused
sibling module `content-fetcher/bundled.ts` (PR 5 of the content-fetcher
modularization stack). content-fetcher.ts drops to ~1,054 LOC.

## What moved (verbatim, no logic change)

- `BundledRef` discriminated union + `parseBundledUrl` + the
  `SAFE_BUNDLED_PACKAGE_PATH` path-safety regex
- `readBundledLocalStorageGuide` and the two localStorage loaders
  (`loadBundledWysiwygPreview`, `loadBundledE2ETest`)
- `loadBundledPackage` (require by relative path) + `loadBundledIndexed`
  (index.json lookup)
- `fetchBundledInteractive` dispatcher

## Reviewer gotcha: require paths

The module moved one directory deeper, so the three
`require('../bundled-interactives/...')` calls become
`require('../../bundled-interactives/...')`. This is the only non-mechanical
edit — verified by the new happy-path tests that actually `require` real
bundled files.

## Barrel stability

None of this cluster is on the public surface — all loaders were
module-private and only `fetchBundledInteractive` is called internally (from
the `bundled:` branch of `fetchInteractiveContent`). It is imported back into
content-fetcher.ts; no barrel change, no consumer touched. The now-unused
`StorageKeys` import is dropped from content-fetcher.ts.

## Tests (+17, gap #6)

New `bundled.test.ts` adds the previously-missing happy paths: real
`require()` resolution for both the package path (title derived from the JSON
module) and the indexed path (title derived from the index.json entry), plus
full `parseBundledUrl` branch coverage (sentinels, safe/unsafe package paths,
traversal rejection, empty, indexed) and the localStorage loaders
(title-present, title-fallback, empty-state). Before this only the error
branches were exercised. The existing
`content-fetcher.bundled-localstorage.test.ts` is untouched and still passes.
